requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Ingestion
kafka-python>=2.0.2

# Storage
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Processing (stream consumer)
confluent-kafka>=2.3.0

# Visualization
streamlit>=1.29.0
pandas>=2.1.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Notebooks & analysis
jupyter>=1.0.0
ipykernel>=6.27.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0

# Utilities
requests>=2.31.0
pytz>=2023.3
numpy>=1.26.0
